[APIS-906] PrepareStmtCache must be maintained until the Connection is closed. (#30)
http://jira.cubrid.org/browse/APIS-906​

When using Prepared Statement Cache using usePreparedStmtCache, preparedStmtCacheSize, and preparedStmtCacheSqlLimitYou can use more queries faster. However, when the statement is closed, the cache is deleted, and the prepare step is performed again when performing the same query after that, resulting in a delay in query execution.To eliminate this delay, the cache should be maintained until the connection is closed.